FinsmartAI/FinSmartBackend/thiru_repo/requirements.txt
langchain>=0.3.10,<0.4.0
langchain-core>=0.3.0
langchain-community>=0.3.10,<0.4.0
langchain-huggingface>=0.1.0
langchain-astradb
astrapy>=2.0.0
streamlit
python-dotenv
transformers
torch
accelerate
sentence-transformers
pypdf
pandas
plotly
langchain-nvidia-ai-endpoints
fastapi
uvicorn
python-multipart
requests